+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2025-03-10
+
+- Add Draupnir as moderation tool. See [Draupnir documentation](https://the-draupnir-project.github.io/draupnir-documentation/) for more information about the project.
+
 ### 2025-02-24
 
 - Add Synapse Service Health Grafana dashboard.